requirements.txt
dotenv
fastapi
loguru
pysaml2
python-multipart
uvicorn 
xmlsec